% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_measures.R
\name{apply_updown}
\alias{apply_updown}
\title{Apply Elevation (updown)}
\usage{
apply_updown(x)
}
\arguments{
\item{x}{Calibrated acceleration data frame.}
}
\value{
Measure column appended to end of calibrated data frame.
}
\description{
Apply Elevation (updown)
}
\examples{
x <- c(0.14268, 0.21757, -0.529, -0.36383)
y <- c(0.26385, 0.27295, 0.29220, 0.79510)
z <- c(0.27722, 0.20296, 0.35092, 0.27459)
calibrated <- data.frame(x, y, z)
calibrated <- apply_updown(calibrated)
}
