% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneabout_functions.R
\name{bouts_decision_tree}
\alias{bouts_decision_tree}
\title{Bouts Decision Tree}
\usage{
bouts_decision_tree(
  bouts,
  SDduration_threshold = 5.7e-05,
  AGSA_threshold = 0.0625,
  running_threshold = 0.407
)
}
\arguments{
\item{bouts}{Aggregated events after non-wear and rest coverage have been assigned.}

\item{SDduration_threshold}{Threshold for sleep classification. Uses the average x,y,z standard deviation over duration metric.}

\item{AGSA_threshold}{Threshold for active classification.}
}
\value{
Classified bouts
}
\description{
Bouts Decision Tree
}
\keyword{internal}
