% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_measures.R
\name{calculate_daily_activity_measures}
\alias{calculate_daily_activity_measures}
\title{Calculate Daily Activity Measures}
\usage{
calculate_daily_activity_measures(
  MPI,
  daily_bouts,
  minimum_valid_hours = 22,
  include_participant_info = FALSE
)
}
\arguments{
\item{MPI}{Measurement Period Information.}

\item{daily_bouts}{Function used to read daily bouts or extract them from the aggregated bouts file}

\item{minimum_valid_hours}{Minimum hours of wear time in a 24-hour day to be considered a valid day for reported measures.}

\item{include_participant_info}{If \code{TRUE}, appends participant information extracted from the bin file to both the activity and sleep measures.}
}
\value{
Activity measures data frame for all days
}
\description{
Calculate Daily Activity Measures
}
\details{
Function that calculates the daily activity measures only
}
\keyword{internal}
