% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPI_create.R
\name{check_MPI}
\alias{check_MPI}
\title{Print out a summary of MPI file(s)}
\usage{
check_MPI(input, print_output = TRUE)
}
\arguments{
\item{input}{Fully qualified path to a MPI file or a directory containing MPI files,
the directory is searched recursively.}

\item{print_output}{Boolean to print MPI summary to console.}
}
\description{
Function to print out a summary of an MPI (Measurement Process
Information) file, or all MPI files found if a directory path is supplied.
}
\examples{
\dontrun{
check_MPI("fully qualified path to a directory structure containing MPI files")
check_MPI("fully qualified path to a MPI file")
}
}
