% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_time_format}
\alias{check_time_format}
\title{Check Time Format}
\usage{
check_time_format(time_str)
}
\arguments{
\item{time_str}{Time string.}
}
\value{
Valid time string or error.
}
\description{
Check Time Format
}
\details{
Internal function to parse and check validity of time string
passed as parameter.
}
\keyword{internal}
