% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coel_json_create.R
\name{create_coel_atoms_json}
\alias{create_coel_atoms_json}
\title{Create COEL Behavioural Atoms JSON}
\usage{
create_coel_atoms_json(data_folder)
}
\arguments{
\item{data_folder}{Folder that contains raw data bin files for which bouts have been created for.}
}
\description{
Creates per-participant rest activity and behavioural bout atoms and a payload
}
