% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{cx_percentile}
\alias{cx_percentile}
\title{Compute Cx percentile}
\usage{
cx_percentile(
  bouts,
  percentile = 95,
  value_col = "StepMean",
  duration_col = "Duration"
)
}
\arguments{
\item{bouts}{A data frame sorted by time containing at minimum a \code{value_col} and \code{duration_col}.}

\item{percentile}{The percentile to compute. Single integer between 0 and 100. Default is 95.}

\item{value_col}{The column name of the variable for which the percentile will be calculated (e.g., \verb{"StepMean}).}

\item{duration_col}{The column name of the duration or weight column (e.g., \code{"Duration"}).}
}
\value{
A single numeric value or \code{NA} if percentile cannot be calculated
}
\description{
Compute Cx percentile
}
\details{
Calculates the xth percentile of cadence for a given percentile.
}
\keyword{internal}
