% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPI_detect.R
\name{detect_transitions}
\alias{detect_transitions}
\title{Detect Transitions}
\usage{
detect_transitions(
  measurements,
  minimum_event_duration = 5,
  x_cpt_penalty = 18,
  y_cpt_penalty = 25,
  z_cpt_penalty = 16,
  cut_time_24hr = "15:00"
)
}
\arguments{
\item{measurements}{The downsampled measurements to calculate changepoint transitions.}

\item{minimum_event_duration}{The minimum interval between changepoint transitions.}

\item{x_cpt_penalty}{The manual penalty value applied in the PELT changepoint
algorithm for the x axis, see \code{\link[changepoint]{cpt.var}}.}

\item{y_cpt_penalty}{The manual penalty value applied in the PELT changepoint
algorithm for the y axis, see \code{\link[changepoint]{cpt.var}}.}

\item{z_cpt_penalty}{The manual penalty value applied in the PELT changepoint
algorithm for the z axis, see \code{\link[changepoint]{cpt.var}}.}

\item{cut_time_24hr}{Time in 24h to split days up by.}
}
\value{
List of time and index of each transition.
}
\description{
Detect Transitions
}
\details{
Function to detect mean and variance changepoints in 1Hz acceleration data from a bin file.
}
