% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expected_last_timestamp}
\alias{expected_last_timestamp}
\title{Calculate Expected Last Timestamp}
\usage{
expected_last_timestamp(measurement_frequency, page_count, first_timestamp)
}
\arguments{
\item{measurement_frequency}{Measurement frequency of recorded data in the bin file.}

\item{page_count}{Total number of pages in the bin file.}

\item{first_timestamp}{Unix timestamp of the first page of recorded data in the bin file.}
}
\value{
Expected last timestamp
}
\description{
Calculate Expected Last Timestamp
}
\details{
Calculates the expected last timestamp of a file based on measurement frequency,
page count and first timestamp of file.
}
\keyword{internal}
