% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneabout_functions.R
\name{find_rest_intervals}
\alias{find_rest_intervals}
\title{Determine Daily Recommended Primary Rest Interval}
\usage{
find_rest_intervals(
  start_expansion_percent,
  end_expansion_percent,
  cut_time_24hr,
  MPI
)
}
\arguments{
\item{start_expansion_percent}{The percentage expansion, based on duration, of a still period's start time to identify overlaps with near-adjacent still periods.}

\item{end_expansion_percent}{The percentage expansion, based on duration, of a still period's end time to identify overlaps with near-adjacent still periods.}

\item{cut_time_24hr}{Time in 24h to split days up by.}

\item{MPI}{Measurement Period Information.}
}
\description{
Determine Daily Recommended Primary Rest Interval
}
\details{
Twice the expansion before than after.
}
\keyword{internal}
