% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_measures.R
\name{folder_measures}
\alias{folder_measures}
\title{Daily Measures for Data Folder}
\usage{
folder_measures(
  data_folder,
  measure_type = c("activity", "sleep", "both"),
  minimum_valid_hours = 22,
  include_participant_info = FALSE,
  identifier_mapping_record = NULL
)
}
\arguments{
\item{data_folder}{Folder that contains raw data bin files to process or path to single bin file.}

\item{measure_type}{The type of daily measure summary to generate.}

\item{minimum_valid_hours}{Minimum hours of wear time in a 24-hour day to be considered a valid day for reported measures.}

\item{include_participant_info}{If \code{TRUE}, appends participant information extracted from the bin file to both the activity and sleep measures.}

\item{identifier_mapping_record}{File path of the exported identifier_mapping_record CSV file.}
}
\value{
Daily aggregated activity measures (.csv)
}
\description{
Daily Measures for Data Folder
}
\details{
Wrapper to process and output daily activity measures aggregate for all bin files in the data folder
}
