% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{fx_count}
\alias{fx_count}
\title{Compute Fx metrics}
\usage{
fx_count(bouts, AGSA_threshold = 0.0625, duration = 3600)
}
\arguments{
\item{bouts}{A bouts data frame sorted by time containing at minimum the TimeUTC and AGSAMean columns.}

\item{AGSA_threshold}{Threshold for active classification in g. Default is 0.0625 (62.5mg).}

\item{duration}{Duration of active time in seconds. Default is 3600 seconds (60 minutes).}
}
\value{
A list containing the maximum and minimum bout counts and their
corresponding start times for the supplied bouts data at a given window.
}
\description{
Compute Fx metrics
}
\details{
Calculates frequency of bouts for each x duration window.
}
\keyword{internal}
