% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneabout.R
\name{geneabout}
\alias{geneabout}
\title{GENEAbout}
\usage{
geneabout(data_folder = data_folder, control = list())
}
\arguments{
\item{data_folder}{Folder that contains raw data bin files to process or path to single bin file.}

\item{control}{Named list of optional settings for geneacore functions.}
}
\description{
GENEAbout function which performs the following tasks for each available day:
\itemize{
\item Assign bout classifications to events only
\item Reports aggregated daily activity and sleep measures
\item Creates COEL v2.0 Behavioural Atoms JSON
}
}
\details{
Control options (all optional):
\describe{
\item{timer}{Logical. Default \code{FALSE}. Print elapsed times of each processing step.}
\item{minimum_valid_hours}{Integer. Default 22.
Minimum hours of wear time in a 24-hour day to be considered a valid day for reported measures.}
\item{save_daily_bouts}{Logical. Default \code{FALSE}.
If \code{TRUE},daily bouts are saved as RDS files and CSV files (if \code{output_csv} is set to \code{TRUE}).
If \code{FALSE}, only a single RDS file containing bouts for all processed days is saved.}
\item{coel_json}{Logical. Default \code{FALSE}. Exports bouts as COEL behavioural atoms in JSON format.
For each bin file in the data folder, behavioural bout atoms and rest activity atoms are produced.
Additionally, aggregated JSON files are created in the \verb{Batch JSON outputs} subfolder.}
\item{identifier_mapping_record}{String. Default \code{NULL}.
If a file path to the exported \code{identifier_mapping_record.csv} is provided,
encrypted participant IDs in the activity and sleep measures will be replaced
with their corresponding mapped identifiers.}
}
}
\examples{
\dontrun{
controls <- list(minimum_valid_hours = 4, timer = TRUE, coel_json = TRUE)
geneabout("path/to/folder", controls)
}

}
