% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneacore_part1.R
\name{geneacore_part1}
\alias{geneacore_part1}
\title{GENEAcore Part 1}
\usage{
geneacore_part1(data_folder = data_folder, control = list())
}
\arguments{
\item{data_folder}{Folder that contains raw data bin files to process or path to single bin file.}

\item{control}{Named list of optional settings for geneacore functions.}
}
\value{
Measurement Period Information (.rds, .json) and 1Hz downsampled data (.rds)
}
\description{
GENEAcore function which performs the following tasks:
\itemize{
\item Checks and reads the data file and creates the Measurement Period Information (MPI)
\item Downsamples the file to 1Hz and detects periods of non movement
\item Calculates auto calibration parameters for the device
}
}
\details{
Control options (all optional):
\describe{
\item{cut_time_24hr}{String. Default "15:00". Time in 24h to split days up by.}
\item{timer}{Logical. Default \code{FALSE}. Print elapsed times of each processing step.}
}
}
\examples{
\dontrun{
controls <- list(output_csv = TRUE, required_processing_hours = 4, timer = TRUE)
geneacore_part1("path/to/folder", controls)
}

}
