% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneacore_part2.R
\name{geneacore_part2}
\alias{geneacore_part2}
\title{GENEAcore Part 2}
\usage{
geneacore_part2(data_folder = data_folder, control = list())
}
\arguments{
\item{data_folder}{Folder that contains raw data bin files to process or path to single bin file.}

\item{control}{Named list of optional settings for geneacore functions.}
}
\value{
Daily raw data sample of x, y, z, Light, Button, Temperature and Voltage (.rds)
}
\description{
GENEAcore function which performs the following tasks for each valid day:
\itemize{
\item Samples daily raw bin file data and saves as daily RDS files
\item Reads and saves button press timestamps to the Measurement Period Information (MPI)
}
}
\details{
Control options (all optional):
\describe{
\item{required_processing_hours}{Integer. Default 0. Hours of wear time in a 24-hour day to be considered a valid day to be processed.}
\item{timer}{Logical. Default \code{FALSE}. Print elapsed times of each processing step.}
}
}
\examples{
\dontrun{
controls <- list(output_csv = TRUE, required_processing_hours = 4, timer = TRUE)
geneacore_part2("path/to/folder", controls)
}

}
