% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneacore_part3.R
\name{geneacore_part3}
\alias{geneacore_part3}
\title{GENEAcore Part 3}
\usage{
geneacore_part3(data_folder = data_folder, control = list())
}
\arguments{
\item{data_folder}{Folder that contains raw data bin files to process or path to single bin file.}

\item{control}{Named list of optional settings for geneacore functions.}
}
\description{
GENEAcore function which performs the following tasks for each valid day:
\itemize{
\item Applies calibration parameters to raw data
\item Identifies transitions for events processing only
\item Performs aggregation by event and/or epoch
\item Calculate steps, if option selected
\item Calculate non-wear coverage and rest coverage of events and/or epochs
}
}
\details{
Control options (all optional):
\describe{
\item{output_epochs}{Logical. Default \code{FALSE}. Create epoch outputs.}
\item{epoch_duration}{Integer. Default 1. Specify duration of fixed epochs.}
\item{output_events}{Logical. Default \code{TRUE}. Create event outputs.}
\item{output_steps}{Logical. Default \code{FALSE}. Calculate step counts and stepping rate during epoch processing.
Steps are always calculated during events processing.}
\item{output_csv}{Logical. Default \code{FALSE}. Allows CSV output to be saved during epoch and event processing.}
\item{timer}{Logical. Default \code{FALSE}. Print elapsed times of each processing step.}
\item{multisession}{Integer vector of length 2. Default \code{c(1,1)}.
Controls day split when running function across multiple R sessions.
The first value is the current session number and the second is the total number of sessions.
The default processes all days in a single session.}
\item{required_processing_hours}{Integer. Default 0. Hours of wear time in a 24-hour day to be considered a valid day to be processed.}
}
}
\examples{
\dontrun{
controls <- list(output_csv = TRUE, required_processing_hours = 4, timer = TRUE)
geneacore_part3("path/to/folder", controls)
}

}
