% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneacore_part4.R
\name{geneacore_part4}
\alias{geneacore_part4}
\title{GENEAcore Part 4}
\usage{
geneacore_part4(data_folder = data_folder, control = list())
}
\arguments{
\item{data_folder}{Folder that contains raw data bin files to process or path to single bin file.}

\item{control}{Named list of optional settings for geneacore functions.}
}
\value{
Consolidated epochs and/or bouts output (.rds/.csv)
}
\description{
GENEAcore function which performs the following tasks:
\itemize{
\item Consolidates daily epochs or bouts outputs into a single file
}
}
\details{
Control options (all optional):
\describe{
\item{output_epochs}{Logical. Default \code{TRUE}. Create epoch outputs.}
\item{epoch_duration}{Integer. Default 1. Specify duration of fixed epochs.}
\item{output_events}{Logical. Default \code{TRUE}. Create event outputs.}
\item{output_csv}{Logical. Default \code{FALSE}. Allows CSV output to be saved during epoch and event processing.}
}
}
\examples{
\dontrun{
controls <- list(output_csv = TRUE, minimum_valid_hours = 4, timer = TRUE)
geneacore_part4("path/to/folder", controls)
}

}
