% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_decimal_places}
\alias{get_decimal_places}
\title{Get Decimal Places}
\usage{
get_decimal_places(column)
}
\arguments{
\item{column}{Aggregated data frame column.}
}
\value{
Decimal place integer.
}
\description{
Get Decimal Places
}
\details{
Internal function to determine the number of decimal places based on column name.
}
\keyword{internal}
