% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_events}
\alias{get_events}
\title{Get events from transitions for aggregation and step counter}
\usage{
get_events(time_series, transitions, sample_frequency)
}
\arguments{
\item{time_series}{Data frame that transitions are for}

\item{transitions}{Returned output of detect_transitions().}

\item{sample_frequency}{Measurement frequency of data.}
}
\description{
Get events from transitions for aggregation and step counter
}
\keyword{internal}
