% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPI_create.R
\name{get_UniqueBinFileIdentifier.dir}
\alias{get_UniqueBinFileIdentifier.dir}
\title{Generate Unique Bin File Identifier for each bin file within a directory path}
\usage{
get_UniqueBinFileIdentifier.dir(binfile_directory)
}
\arguments{
\item{binfile_directory}{Path to the file to be processed}
}
\value{
Named vector, the vector name is the filename and the value the identifier.
}
\description{
Generate Unique Bin File Identifier for each bin file within a directory path
}
\details{
Function to create a UniqueBinFileIdentifier from a GENEActiv bin file.
}
\examples{
binfile_directory <- system.file("extdata", package = "GENEAcore")
UniqueBinFileIdentifiers <- get_UniqueBinFileIdentifier.dir(binfile_directory)
}
\keyword{internal}
