% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPI_create.R
\name{get_UniqueBinFileIdentifier.filepath}
\alias{get_UniqueBinFileIdentifier.filepath}
\title{Generate Unique Bin File Identifier from a file path}
\usage{
get_UniqueBinFileIdentifier.filepath(binfile_path)
}
\arguments{
\item{binfile_path}{Path to the file to be processed}
}
\value{
Single string identifier.
}
\description{
Generate Unique Bin File Identifier from a file path
}
\details{
Function to create a UniqueBinFileIdentifier from a GENEActiv bin file.
}
\examples{
binfile_path <- system.file("extdata/10Hz_calibration_file_20Nov25.bin", package = "GENEAcore")
UniqueBinFileIdentifier <- get_UniqueBinFileIdentifier.filepath(binfile_path)
}
\keyword{internal}
