% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{identifier_mapping}
\alias{identifier_mapping}
\title{Map Study Management Hub Identifiers}
\usage{
identifier_mapping(x, cols_to_map, identifier_mapping_record)
}
\arguments{
\item{x}{Data frame with ID columns to map.}

\item{cols_to_map}{Vector of column names to map ID to identifier.}

\item{identifier_mapping_record}{File path of the exported identifier_mapping_record CSV file.}
}
\description{
Map Study Management Hub Identifiers
}
\keyword{internal}
