% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MPI_toJSON}
\alias{MPI_toJSON}
\title{Write MPI to JSON}
\usage{
MPI_toJSON(MPI, mpi_path)
}
\arguments{
\item{MPI}{Measurement Period Information.}

\item{mpi_path}{File path to MPI RDS.}
}
\value{
MPI JSON file written to the same directory as the MPI RDS.
}
\description{
Write MPI to JSON
}
\details{
Restructures the MPI and writes it out to a JSON for ActivInsights internal use.
}
\keyword{internal}
