% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{mx_lx_window}
\alias{mx_lx_window}
\title{Compute Mx and Lx metrics}
\usage{
mx_lx_window(bouts, window, low = TRUE)
}
\arguments{
\item{bouts}{A data frame sorted by time containing at minimum the columns TimeUTC, Duration and AGSAMean.}

\item{window}{Window length in seconds.}

\item{low}{Set to TRUE for both the maximum (Mx) and minimum (Lx) intensities
are returned. If FALSE, only the maximum (Mx) is returned.}
}
\value{
A list containing the maximum and minimum mean intensity and their
corresponding start times for the supplied bouts data at a given window.
}
\description{
Compute Mx and Lx metrics
}
\details{
Calculates intensity for each x duration window.
}
\keyword{internal}
