% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPI_detect.R
\name{new_cut_times}
\alias{new_cut_times}
\title{New Cut Times}
\usage{
new_cut_times(df)
}
\arguments{
\item{df}{Cut Times data frame.}
}
\value{
Data frame with added cut times.
}
\description{
New Cut Times
}
\details{
Add the timestamps, indexes and day numbers of the cut times and their ends.
}
\examples{
CutTimes_df <- data.frame(
  time = c(1731421000, 1731421100, 1731421362, 1731421480, 1731421525),
  index = c(56, 1, 230, 1, 400), day = c(1, 2, 2, 3, 3)
)
CutTimes_df <- new_cut_times(CutTimes_df)
}
