% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneabout_functions.R
\name{nonwear_rest_coverage}
\alias{nonwear_rest_coverage}
\title{Non-wear and Rest Coverage for Epochs/Events}
\usage{
nonwear_rest_coverage(
  MPI,
  aggregated_data,
  start_time = NULL,
  end_time = NULL,
  nonwear_day
)
}
\arguments{
\item{MPI}{Measurement Period Information.}

\item{aggregated_data}{Aggregated epochs or events.}

\item{start_time}{Time stamp to start the coverage assignment, defaults to first time stamp of aggregated data object passed.}

\item{end_time}{Time stamp to end the coverage assignment, defaults to end time stamp of aggregated data object passed.}

\item{nonwear_day}{Data frame of non-wear periods and their durations.}
}
\description{
Non-wear and Rest Coverage for Epochs/Events
}
\keyword{internal}
