% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reorder_df}
\alias{reorder_df}
\title{Reorder Data Frame}
\usage{
reorder_df(df)
}
\arguments{
\item{df}{Epochs or events data frame.}
}
\value{
Reordered data frame.
}
\description{
Reorder Data Frame
}
\details{
Internal function to remove the epoch_number or event_number
columns and move duration column forward.
}
\keyword{internal}
