% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{round_columns}
\alias{round_columns}
\title{Round Columns}
\usage{
round_columns(df)
}
\arguments{
\item{df}{Aggregated data frame.}
}
\value{
Epochs or events data frame with rounded columns.
}
\description{
Round Columns
}
\details{
Internal function to round columns based on column names or
variability of data in column.
}
\keyword{internal}
