% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_measures.R
\name{single_file_measures}
\alias{single_file_measures}
\title{Daily Measures for Single Bin File}
\usage{
single_file_measures(
  output_folder,
  UniqueBinFileIdentifier,
  measure_type = c("activity", "sleep", "both"),
  minimum_valid_hours = 22,
  include_participant_info = FALSE,
  identifier_mapping_record = NULL
)
}
\arguments{
\item{output_folder}{Path to the folder containing GENEAcore run outputs and Measurement Period Information (MPI) files.}

\item{UniqueBinFileIdentifier}{The Unique Bin File Identifier of a bin file.}

\item{measure_type}{The type of daily measure summary to generate.}

\item{minimum_valid_hours}{Minimum hours of wear time in a 24-hour day to be considered a valid day for reported measures.}

\item{include_participant_info}{If \code{TRUE}, appends participant information extracted from the bin file to both the activity and sleep measures.}

\item{identifier_mapping_record}{File path of the exported identifier_mapping_record CSV file.}
}
\value{
Daily aggregated activity measures (.csv)
}
\description{
Daily Measures for Single Bin File
}
\details{
Wrapper to process and output daily activity measures aggregate for a single bin file
}
\keyword{internal}
