% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_counter.R
\name{step_counter}
\alias{step_counter}
\alias{stepCounter}
\title{Step Counter}
\usage{
step_counter(
  step_data,
  sample_frequency = 100,
  filter_order = 2,
  boundaries = c(0.5, 5),
  Rp = 3,
  step_hysteresis = 0.05,
  fun = c("GENEAcount", "mean", "sd", "stepdiff")
)

stepCounter(...)
}
\arguments{
\item{step_data}{The data to use for calculating the steps. This should be a vector of acceleration values.}

\item{sample_frequency}{The sampling frequency of the data, in hertz,
when calculating the step number (default 100).}

\item{filter_order}{single integer, order of the Chebyshev bandpass filter,
passed to argument n of \code{\link[signal]{cheby1}}.}

\item{boundaries}{length 2 numeric vector specifying lower and upper bounds
of Chebychev filter (default \code{c(0.5, 5)} Hz),
passed to argument W of \code{\link[signal]{butter}} or \code{\link[signal]{cheby1}}.}

\item{Rp}{the decibel level that the cheby filter takes, see \code{\link[signal]{cheby1}}.}

\item{step_hysteresis}{The hysteresis applied after zero crossing of the bandpass filtered y-axis signal.}

\item{fun}{character vector naming functions by which to summarize steps.
"count" is an internally implemented summarizing function that returns step count.}

\item{...}{Additional arguments passed to internal aggregation functions.}
}
\value{
Returns a vector with length fun.
}
\description{
Function to calculate the number and variance of the steps in the data.
}
\examples{
d1 <- sin(seq(0.1, 100, 0.1)) / 2 + rnorm(1000) / 10 + 1
Steps4 <- step_counter(d1)
length(Steps4)
mean(Steps4)
sd(Steps4)
plot(Steps4)
}
