% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{valid_day_nonwear}
\alias{valid_day_nonwear}
\title{Determine Valid Days}
\usage{
valid_day_nonwear(
  MPI_nonwear,
  date_range,
  required_processing_hours = 0,
  cut_time_24hr = "15:00"
)
}
\arguments{
\item{MPI_nonwear}{MPI non-wear data frame with start times and durations of non-wear events.}

\item{date_range}{Data frame of day border timestamps.}

\item{required_processing_hours}{Hours of wear time in a 24-hour day to be considered a valid day to be processed.}

\item{cut_time_24hr}{Time in 24h to split days up by.}
}
\value{
List of 2 data objects:
\itemize{
\item valid_day provides information on the wear and non-wear duration for each day and its Boolean validity based on \code{minimum_valid_day}.
\item nonwear_day provides the start and end time of each non-wear period and the day in which it occurs.
}
}
\description{
Determine Valid Days
}
\details{
Determine which days of data are valid days to process based on non-wear
calculated in MPI and a minimum valid day duration.
}
\keyword{internal}
