% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_measures.R
\name{apply_all}
\alias{apply_all}
\title{Apply AGSA, ENMO, UpDown and Degrees}
\usage{
apply_all(x)
}
\arguments{
\item{x}{Calibrated acceleration data frame.}
}
\value{
Measures columns appended to end of calibrated data frame.
}
\description{
Single function that combines calculations for acceleration, elevation and rotation measures.
}
\examples{
x <- c(0.14268, 0.21757, -0.529, -0.36383)
y <- c(0.26385, 0.27295, 0.29220, 0.79510)
z <- c(0.27722, 0.20296, 0.35092, 0.27459)
calibrated <- data.frame(x, y, z)
calibrated <- apply_all(calibrated)
}
