% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPI_create.R
\name{binfile_summary}
\alias{binfile_summary}
\title{Bin File Summary}
\usage{
binfile_summary(input, recursive = TRUE)
}
\arguments{
\item{input}{Bin file path.}

\item{recursive}{TRUE applies the operation to all nested elements.}
}
\value{
Data frame of bin file or MPI summary.
}
\description{
Bin File Summary
}
\details{
Wrapper function that calls \code{create_summary} for bin files only.
}
