% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneacore.R
\name{geneacore}
\alias{geneacore}
\title{Main GENEAcore Function}
\usage{
geneacore(
  data_folder = data_folder,
  CutTime24Hr = "15:00",
  output_epochs = TRUE,
  epoch_duration = 1,
  output_events = TRUE,
  output_steps = FALSE,
  output_csv = FALSE,
  timer = FALSE
)
}
\arguments{
\item{data_folder}{Folder that contains raw data bin files to process.}

\item{CutTime24Hr}{Time in 24h to split the days up by.}

\item{output_epochs}{Create epoch outputs.}

\item{epoch_duration}{Specify duration of fixed epochs.}

\item{output_events}{Create event outputs.}

\item{output_steps}{Include step counts and stepping rate outputs.}

\item{output_csv}{Allows CSV output to be saved during epoch and event processing.}

\item{timer}{Print elapsed times of each process.}
}
\value{
RDS and CSV of Measurement Period Information, Epoch measures and Event measures.
}
\description{
Main GENEAcore Function
}
