% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_cut_times}
\alias{get_cut_times}
\title{Get Cut Times}
\usage{
get_cut_times(cut_time_24hr, MPI)
}
\arguments{
\item{cut_time_24hr}{Time in 24h to split the days up by.}

\item{MPI}{Measurement Period Information.}
}
\value{
List of cut times in Unix UTC.
}
\description{
Get Cut Times
}
\details{
Internal function to determine the daily cut times based on 24Hr Cut Time specified.
}
\keyword{internal}
