% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPI_create.R
\name{MPI_summary}
\alias{MPI_summary}
\title{MPI Summary}
\usage{
MPI_summary(input, recursive = TRUE)
}
\arguments{
\item{input}{MPI path - single MPI file or folder of MPI files.}

\item{recursive}{TRUE applies the operation to all nested elements.}
}
\value{
Data frame of MPI summary.
}
\description{
MPI Summary
}
\details{
Wrapper function that calls \code{create_summary} for MPI only.
}
