% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{round_columns}
\alias{round_columns}
\title{Round Columns}
\usage{
round_columns(df)
}
\arguments{
\item{df}{Aggregated data frame.}
}
\value{
Epochs or events data frame with rounded columns.
}
\description{
Round Columns
}
\details{
Internal function to round columns based on column names or
variability of data in column.
}
\examples{
epochs_df <- data.frame(
  "x.mean" = c(0.1111, 0.1222, 0.1333, 0.1444),
  "y.mean" = c(0.2111, 0.2222, 0.2333, 0.2444),
  "light.mean" = c(1.25, 1.73, 1.99, 2.02)
)
epochs_df <- round_columns(epochs_df)
}
\keyword{internal}
